Fix/handle error

+++ server/controller/ProductsController.js
@@ -1,16 +1,23 @@
 const database = require("../db/models");
 
 class ProductsController {
-  static get(req, res, next) {
-    const products = database.Products.findAll();
+  static productExist(id) {
+    const searchById = database.Products.findByPk(id);
+    return searchById;
+  }
+
+  static getAllProducts(req, res, next) {
+    const products = database.Products.findAll({
+      order: [['id', 'ASC']],
+    });
     products
       .then((result) => {
         res.status(200).json(result);
       })
-      .catch((err) => next({ code: 403, err }));
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static getById(req, res, next) {
+  static getProductById(req, res, next) {
     const { id } = req.params;
     const product = database.Products.findOne({
       where: {
@@ -21,29 +28,44 @@ class ProductsController {
       .then((result) => {
         res.status(200).json(result);
       })
-      .catch((err) => next({ code: 403, err }));
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static post(req, res, next) {
+  static postProduct(req, res, next) {
     const { name, price, menu, type, flavor, restaurant } = req.body;
-    const createProduct = database.Products.create({
-      name,
-      price,
-      menu,
-      type,
-      flavor,
-      restaurant
+
+    if (!name || !price || !menu || !type || !restaurant) {
+      return res.status(401).json({ message: "Missing required data." });
+    }
+
+    const createProduct = database.Products.findOrCreate({
+      where: { name },
+      defaults: { name, price, menu, type, flavor, restaurant }
     });
+
     createProduct
       .then((result) => {
-        return res.status(201).json(result);
+        const [productObj, status] = result;
+
+        if (status) {
+          return res.status(201).json(productObj.toJSON());
+        } else {
+          return res.status(403).json({ code: 403, msg: "Email already used" });
+        }
       })
-      .catch(next);
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static update(req, res, next) {
+  static async updateProduct(req, res, next) {
     const { id } = req.params;
     const { price, menu, flavor, restaurant } = req.body;
+    
+    const isRegistred = await ProductsController.productExist(id);
+
+    if (!isRegistred) {
+      return res.status(404).json({ code: 404, message: "Product not found." });
+    }
+
     const updateProduct = database.Products.update(
       { price: price, menu: menu, flavor: flavor, restaurant: restaurant },
       {
@@ -52,25 +74,34 @@ class ProductsController {
         },
       }
     );
+
     updateProduct
       .then((result) => {
         return res.status(200).json(result);
       })
-      .catch(next);
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static delete(req, res, next) {
+  static async deleteProduct(req, res, next) {
     const { id } = req.params;
+
+    const isRegistred = await ProductsController.productExist(id);
+
+    if (!isRegistred) {
+      return res.status(404).json({ code: 404, message: "Product not found." });
+    }
+
     const product = database.Products.destroy({
       where: {
         id: Number(id),
       },
     });
+
     product
       .then((result) => {
         return res.status(200).json(result);
       })
-      .catch(next);
+      .catch((err) => next({ code: 400, err }));
   }
 }
