+++ server/controller/UserController.js
@@ -1,19 +1,26 @@
 const database = require("../db/models");
 
 class UserController {
-  static get(req, res, next) {
+  static userExist(id) {
+    const searchById = database.Users.findByPk(id);
+    return searchById;
+  }
+
+  static getAllUsers(req, res, next) {
     const users = database.Users.findAll({
+      order: [["id", "ASC"]],
       attributes: { exclude: ["password"] },
     });
     users
       .then((result) => {
         res.status(200).json(result);
       })
-      .catch((err) => next({ code: 403, err }));
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static getById(req, res, next) {
+  static getUserById(req, res, next) {
     const { id } = req.params;
+
     const user = database.Users.findOne(
       { attributes: { exclude: ["password"] } },
       {
@@ -22,32 +29,50 @@ class UserController {
         },
       }
     );
+
     user
       .then((result) => {
         res.status(200).json(result);
       })
-      .catch((err) => next({ code: 403, err }));
+      .catch((err) => next({ code: 404, err }));
   }
 
-  static post(req, res, next) {
+  static postUser(req, res, next) {
     const { name, email, password, role, restaurant } = req.body;
-    const createUser = database.Users.create({
-      name,
-      email,
-      password,
-      role,
-      restaurant,
+
+    if (!name || !email || !password || !role || !restaurant) {
+      return res.status(401).json({ message: "Missing required data." });
+    }
+
+    const createUser = database.Users.findOrCreate({
+      where: { email },
+      defaults: { name, email, password, role, restaurant },
     });
+
     createUser
       .then((result) => {
-        return res.status(201).json(result);
+        const [userObj, status] = result;
+        const { password, ...user } = userObj.toJSON();
+
+        if (status) {
+          return res.status(201).json(user);
+        } else {
+          return res.status(403).json({ code: 403, msg: "Email already used" });
+        }
       })
-      .catch(next);
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static update(req, res, next) {
+  static async updateUser(req, res, next) {
     const { id } = req.params;
     const { name, password, role } = req.body;
+
+    const isRegistred = await UserController.userExist(id);
+
+    if (!isRegistred) {
+      return res.status(404).json({ code: 404, message: "User not found." });
+    }
+
     const updateUser = database.Users.update(
       { name: name, password: password, role: role },
       {
@@ -56,25 +81,34 @@ class UserController {
         },
       }
     );
+
     updateUser
       .then((result) => {
         return res.status(200).json(result);
       })
-      .catch(next);
+      .catch((err) => next({ code: 400, err }));
   }
 
-  static delete(req, res, next) {
+  static async deleteUser(req, res, next) {
     const { id } = req.params;
+
+    const isRegistred = await UserController.userExist(id);
+
+    if (!isRegistred) {
+      return res.status(404).json({ code: 404, message: "User not found." });
+    }
+
     const user = database.Users.destroy({
       where: {
         id: Number(id),
       },
     });
+
     user
       .then((result) => {
         return res.status(200).json(result);
       })
-      .catch(next);
+      .catch((err) => next({ code: 400, err }));
   }
 }
